move to main, readme
