+++ requirements.txt
@@ -1,1 +1,3 @@
 pyorbital
+requests
+certifi
